+++ axelor-business-project/src/main/java/com/axelor/apps/businessproject/service/ProductTaskTemplateServiceImpl.java
@@ -35,16 +35,16 @@ import java.util.List;
 public class ProductTaskTemplateServiceImpl implements ProductTaskTemplateService {
 
   protected ProductTaskTemplateRepository repository;
-  protected TeamTaskBusinessService teamTaskBusinessService;
+  protected TeamTaskBusinessProjectService teamTaskBusinessProjectService;
   protected TeamTaskRepository teamTaskRepository;
 
   @Inject
   public ProductTaskTemplateServiceImpl(
       ProductTaskTemplateRepository repository,
-      TeamTaskBusinessService teamTaskBusinessService,
+      TeamTaskBusinessProjectService teamTaskBusinessProjectService,
       TeamTaskRepository teamTaskRepository) {
     this.repository = repository;
-    this.teamTaskBusinessService = teamTaskBusinessService;
+    this.teamTaskBusinessProjectService = teamTaskBusinessProjectService;
     this.teamTaskRepository = teamTaskRepository;
   }
 
@@ -66,7 +66,8 @@ public class ProductTaskTemplateServiceImpl implements ProductTaskTemplateServic
       while (qtyTmp.signum() > 0) {
         LocalDateTime dateWithDelay = startDate.plusHours(template.getDelayToStart().longValue());
 
-        TeamTask task = teamTaskBusinessService.create(template, project, dateWithDelay, qty);
+        TeamTask task =
+            teamTaskBusinessProjectService.create(template, project, dateWithDelay, qty);
         task.setParentTask(parent);
         task.setProduct(product);
         task.setQuantity(!template.getIsUniqueTaskForMultipleQuantity() ? BigDecimal.ONE : qty);